LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-getStats-timestamp.https-expected.txt


PASS RTCStats.timestamp is expressed as Performance time
PASS RTCRemoteInboundRtpStats.timestamp is expressed as Performance time
